Extract cobra_commander-yarn
