+++ cobra_commander-yarn/.standard.yml
@@ -1,0 +1,3 @@
+# For available configuration options, see:
+#   https://github.com/testdouble/standard
+ruby_version: 2.6
