+++ cobra_commander-yarn/lib/cobra_commander/yarn.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require_relative "yarn/version"
+require_relative "yarn/package"
+require_relative "yarn/workspace"
+
+module CobraCommander
+  module Yarn
+  end
+end
