+++ cobra_commander-yarn/lib/cobra_commander/yarn/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module CobraCommander
+  module Yarn
+    VERSION = "0.1.0"
+  end
+end
