+++ cobra_commander-yarn/sig/cobra_commander/yarn.rbs
@@ -1,0 +1,6 @@
+module CobraCommander
+  module Yarn
+    VERSION: String
+    # See the writing guide of rbs: https://github.com/ruby/rbs#guides
+  end
+end
