+++ cobra_commander-yarn/spec/cobra_commander/yarn_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+RSpec.describe CobraCommander::Yarn do
+  it "has a version number" do
+    expect(CobraCommander::Yarn::VERSION).not_to be nil
+  end
+
+  it "does something useful" do
+    expect(false).to eq(true)
+  end
+end
