+++ cobra_commander/Gemfile
@@ -8,3 +8,4 @@ bundler_version = ENV.fetch("BUNDLER_VERSION", "2")
 gem "bundler", "~> #{bundler_version}"
 
 gem "cobra_commander-ruby", path: "../"
+gem "cobra_commander-yarn", path: "../"
