--- cobra_commander/lib/cobra_commander/dependencies/yarn/package.rb
@@ -1,21 +1,0 @@
-# frozen_string_literal: true
-
-module CobraCommander
-  module Dependencies
-    class Yarn::Package
-      attr_reader :path, :name, :dependencies
-
-      def initialize(path:, dependencies:, name: nil)
-        @path = path
-        @name = untag(name)
-        @dependencies = dependencies.map { |dep| untag(dep) }
-      end
-
-    private
-
-      def untag(name)
-        name&.gsub(%r{^@[\w-]+/}, "")
-      end
-    end
-  end
-end
